requirements.txt
python-jose[cryptography]>=3.3.0
pydantic>=1.9.0
requests>=2.28.0
aiosql>=9.0
sqlalchemy>=1.4.0
cryptography>=36.0.0
flask>=2.0.0
fastapi>=0.75.0
uvicorn[standard]>=0.17.0
gunicorn>=20.1.0
python-dotenv>=0.20.0